Libertalia: Winds of Galecrest
This game requires one player to shuffle his deck to generate 18 random
cards.  Instead, this app chooses those 18 cards so that each player can
keep his deck in order.

The app uses a cookie to remember the order in case the web page is
reloaded.  The cookie expires after a week.